requirements.txt
asgiref==3.2.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.3
idna==2.10
Pillow==8.0.1
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
